Revert "Gs 248 menu load game"
