+++ src/main/java/org/project/Player.java
@@ -1,7 +1,5 @@
 package org.project;
 
-import org.json.simple.JSONObject;
-
 import javax.imageio.ImageIO;
 import java.awt.*;
 import java.awt.image.BufferedImage;
@@ -295,24 +293,5 @@ public class Player extends Entity {
       }
       gp.ui.showMessage("Monster, RUN!");
     }
-  }
-
-  /**
-   * Sets Player variables from JSON playerData
-   * @param playerData, JSONObject from SaveState
-   */
-  public void loadPlayerData(JSONObject playerData) {
-    if (playerData == null) {
-      throw new NullPointerException("Player object is null.");
-    }
-    this.worldX = ((Long) playerData.get("worldX")).intValue();
-    this.worldY = ((Long) playerData.get("worldY")).intValue();
-    this.speed = ((Long) playerData.get("speed")).intValue();
-    int directionOrdinal = ((Long) playerData.get("direction")).intValue();
-    this.direction = values()[directionOrdinal];
-    this.spriteCounter = ((Long) playerData.get("spriteCounter")).intValue();
-    this.spriteNum = ((Long) playerData.get("spriteNum")).intValue();
-    this.currentLives= ((Long) playerData.get("lives")).intValue();
-    this.currentRubies = ((Long) playerData.get("rubies")).intValue();
   }
 }
