+++ src/main/java/org/project/SaveStateHandler.java
@@ -10,15 +10,14 @@ import java.io.*;
  */
 public class SaveStateHandler {
   private String dirPath = "save/";
-  private String extension = "save_2.json";
-  private String username;
-  private String pathName = dirPath + username + extension;
+  private String fileName = "save_2.json";
+  private String pathName = dirPath + fileName;
 
   /**
    * Stores save data as a JSON file in save directory.
    * @param saveState, current saveState of the game
    */
-  public void save(SaveState saveState) {
+  public void save(SaveState saveState) throws IOException {
     JSONObject jsonSave = new JSONObject();
     jsonSave.put("playerData", saveState.getPlayerData());
     jsonSave.put("gamePanelData", saveState.getGamePanelData());
@@ -29,19 +28,17 @@ public class SaveStateHandler {
       fileWriter.write(jsonSave.toJSONString());
       fileWriter.close();
     } catch (IOException e) {
-      throw new RuntimeException("Could not create or write to file at: " + pathName, e);
+      throw new RuntimeException("Could not create or write to file at: " + dirPath + fileName, e);
     }
   }
 
   /**
-   * Loads save data JSON file from directory and returns saveState.
-   * @param username a String, name of the file
+   * Loads save data JSON file from directory.
+   * @param fileName a String, name of the file
    * @throws FileNotFoundException
-   * @return SaveState object
    */
-  public SaveState load(String username) throws FileNotFoundException {
-    this.username = username;
-    File saveFile = new File(pathName);
+  public void load(String fileName) throws FileNotFoundException {
+    File saveFile = new File(fileName);
     JSONObject jsonSave;
     try {
       FileReader fileReader = new FileReader(saveFile);
@@ -49,16 +46,7 @@ public class SaveStateHandler {
     } catch (FileNotFoundException e) {
       throw new FileNotFoundException("Cannot locate file:" + pathName);
     }
-    SaveState saveState = SaveState.getInstance();
+    SaveState saveState = new SaveState();
     saveState.setSaveState(jsonSave);
-    return saveState;
-  }
-
-  /**
-   * Sets username.
-   * @param username a String, used as file name of save data
-   */
-  public void setUsername(String username) {
-    this.username = username;
   }
 }
